Fixed host-list command-line parameter on Android.
There is a case, when IPv6 socket is created, but IPv4 ip-address
is specified as allowed host. It is possible in this case that
getaddrinfo() returns only IPv4 addrinfo on Android.
Solution is to convert IPv4 addrinfo to IPv4-mapped-IPv6 addrinfo.
This is done manually, because AI_V4MAPPED flag is not supported by
bionic libc.
The reverse case is not supported, because no generic IPv6-address
can be converted to IPv4-address.

Change-Id: I91d7c397179ce4a1053e44c6d0850dffc322c6ed